Added delay after delete operation
While running large scale testing we noticed that the same redis random object was being picked for both DELETE and GET. When the DELETE finished first the GET would generate a failure message.  Through trial-and-error by injected a 25 millisecond delay into the DELETE function seemed to eliminate the failures.  With a 20 ms delay we still see ~3% failures. Unfortunately this delay is also reflected in the RPS and response time measurements.